+++ packages/angular-workspace/nimble-angular/anchor-step/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-anchor-step.directive';
+export * from './nimble-anchor-step.module';