+++ packages/react-workspace/nimble-react/src/stepper/index.ts
@@ -1,0 +1,6 @@
+import { Stepper, stepperTag } from '@ni/nimble-components/dist/esm/stepper';
+import { wrap } from '../utilities/react-wrapper';
+
+export { stepperTag };
+export { type Stepper };
+export const NimbleStepper = wrap(Stepper);